package/requirements.txt
cloudshell-automation-api>=2021.2.0,<2021.2.1
cloudshell-shell-core>=5.0.0,<6.0.0
cloudshell-core>=2.2.0,<2.3.0
pywinrm>=0.2.2
paramiko==2.7.2
requests>=2.16.2
jsonpickle